# src/main/resources/openapi.json
{
	"openapi": "3.0.2",
	"info": {
		"title": "API Title",
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://localhost:8080"
		}
	],
	"paths": {
		"/kunden/listeVonAllen": {
			"get": {
				"summary": "Alle Kunden abrufen",
				"operationId": "getKunden",
				"responses": {
					"200": {
						"description": "Liste aller Kunden",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Kunde"
									}
								}
							}
						}
					}
				}
			}
		},
		"/kunden/{id}": {
			"get": {
				"summary": "Kunde nach ID abrufen",
				"operationId": "getKundeById",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						},
						"description": "ID des Kunden"
					}
				],
				"responses": {
					"200": {
						"description": "Details des Kunden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Kunde"
								}
							}
						}
					},
					"404": {
						"description": "Kunde nicht gefunden"
					}
				}
			},
			"put": {
				"summary": "Kunde aktualisieren",
				"operationId": "updateKunde",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						},
						"description": "ID des Kunden"
					}
				],
				"requestBody": {
					"description": "Aktualisierte Kundendaten",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Kunde"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Aktualisierter Kunde",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Kunde"
								}
							}
						}
					},
					"404": {
						"description": "Kunde nicht gefunden"
					}
				}
			},
			"delete": {
				"summary": "Kunde löschen",
				"operationId": "deleteKunde",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						},
						"description": "ID des Kunden"
					}
				],
				"responses": {
					"200": {
						"description": "Kunde gelöscht",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"example": "Kunde gelöscht!"
								}
							}
						}
					},
					"404": {
						"description": "Kunde nicht gefunden"
					}
				}
			}
		},
		"/kunden": {
			"post": {
				"summary": "Neuen Kunden hinzufügen",
				"operationId": "addKunde",
				"requestBody": {
					"description": "Neuer Kunde",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Kunde"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Hinzugefügter Kunde",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Kunde"
								}
							}
						}
					}
				}
			}
		},
		"/artikel/listeVonAllen": {
			"get": {
				"summary": "Alle Artikel abrufen",
				"operationId": "getArtikel",
				"responses": {
					"200": {
						"description": "Liste aller Artikel",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Artikel"
									}
								}
							}
						}
					}
				}
			}
		},
		"/artikel/{id}": {
			"get": {
				"summary": "Artikel nach ID abrufen",
				"operationId": "getArtikelById",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						},
						"description": "ID des Artikels"
					}
				],
				"responses": {
					"200": {
						"description": "Details des Artikels",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Artikel"
								}
							}
						}
					},
					"404": {
						"description": "Artikel nicht gefunden"
					}
				}
			},
			"put": {
				"summary": "Artikel aktualisieren",
				"operationId": "updateArtikel",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						},
						"description": "ID des Artikels"
					}
				],
				"requestBody": {
					"description": "Aktualisierte Artikeldaten",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Artikel"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Aktualisierter Artikel",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Artikel"
								}
							}
						}
					},
					"404": {
						"description": "Artikel nicht gefunden"
					}
				}
			},
			"delete": {
				"summary": "Artikel löschen",
				"operationId": "deleteArtikel",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						},
						"description": "ID des Artikels"
					}
				],
				"responses": {
					"200": {
						"description": "Artikel gelöscht",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"example": "Artikel gelöscht!"
								}
							}
						}
					},
					"404": {
						"description": "Artikel nicht gefunden"
					}
				}
			}
		},
		"/artikel": {
			"post": {
				"summary": "Neuen Artikel hinzufügen",
				"operationId": "addArtikelBestellung",
				"requestBody": {
					"description": "Neuer Artikel",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Artikel"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Hinzugefügter Artikel",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Artikel"
								}
							}
						}
					}
				}
			}
		},
		"/bestellung/artikel": {
			"get": {
				"summary": "Alle Artikel abrufen",
				"operationId": "getArtikelList",
				"responses": {
					"200": {
						"description": "Liste aller Artikel",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Artikel"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Neuen Artikel hinzufügen",
				"operationId": "addArtikel",
				"requestBody": {
					"description": "Neuer Artikel",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Artikel"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Hinzugefügter Artikel",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Artikel"
								}
							}
						}
					}
				}
			}
		},
		"/bestellung/kunden": {
			"get": {
				"summary": "Alle Kunden abrufen",
				"operationId": "getKundeList",
				"responses": {
					"200": {
						"description": "Liste aller Kunden",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Kunde"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Neuen Kunden hinzufügen",
				"operationId": "addKundeBestellung",
				"requestBody": {
					"description": "Neuer Kunde",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Kunde"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Hinzugefügter Kunde",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Kunde"
								}
							}
						}
					}
				}
			}
		},
		"/bestellung/kunden/{kundeId}/artikel": {
			"post": {
				"summary": "Artikel zu einem Kunden hinzufügen",
				"operationId": "addArtikelToKunde",
				"parameters": [
					{
						"name": "kundeId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						},
						"description": "ID des Kunden"
					}
				],
				"requestBody": {
					"description": "Liste von Artikeln",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Artikel"
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Hinzugefügte Artikel",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Artikel"
									}
								}
							}
						}
					},
					"404": {
						"description": "Kunde nicht gefunden"
					}
				}
			},
			"get": {
				"summary": "Artikel für einen bestimmten Kunden abrufen",
				"operationId": "getArtikelForKunde",
				"parameters": [
					{
						"name": "kundeId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						},
						"description": "ID des Kunden"
					}
				],
				"responses": {
					"200": {
						"description": "Liste der Artikel für den Kunden",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Artikel"
									}
								}
							}
						}
					},
					"404": {
						"description": "Kunde nicht gefunden"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Kunde": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"example": 1
					},
					"name": {
						"type": "string",
						"example": "Max Mustermann"
					},
					"telefonNr": {
						"type": "string",
						"example": "123456789"
					},
					"artikelList": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Artikel"
						}
					}
				}
			},
			"Artikel": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"example": 101
					},
					"name": {
						"type": "string",
						"example": "Artikel A"
					},
					"preis": {
						"type": "number",
						"format": "double",
						"example": 19.99
					}
				}
			}
		}
	}
}